fix lints

+++ x/auth/ante/setup.go
@@ -53,7 +53,7 @@ func (sud SetUpContextDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, _ bool, 
 
 	// If there exists a maximum block gas limit, we must ensure that the tx
 	// does not exceed it.
-	if maxGas > 0 && gasTx.GetGas() > uint64(maxGas) {
+	if maxGas > 0 && gasTx.GetGas() > maxGas {
 		return newCtx, errorsmod.Wrapf(sdkerrors.ErrInvalidGasLimit, "tx gas limit %d exceeds block max gas %d", gasTx.GetGas(), maxGas)
 	}
